pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seymourlib"
version = "0.1.0"
description = "Async Python client for Seymour masking screens (direct serial or IP2SL)"
authors = [{name = "Richard Berg", email = "ars@richardberg.net"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=3.4",
    "attrs>=25.4.0",
    "tenacity>=8.0",
    "async-typer>=0.1.10",
]

[project.optional-dependencies]
serial = [
    "pyserial_asyncio>=0.6",
]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2",
    "coverage>=7",
    "mypy>=1.10",
    "ruff>=0.6",
    "types-setuptools",
    "pyserial_asyncio>=0.6",
]
lint = [
    "ruff>=0.6",
]
typecheck = [
    "mypy>=1.10",
]
test = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2",
    "coverage>=7",
    "pyserial_asyncio>=0.6",
]

[project.scripts]
seymourctl = "seymourlib.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true
unsafe-fixes = false
lint.select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # warnings
    "I",   # import sorting
    "UP",  # pyupgrade
    "B",   # bugbear
    "ASYNC", # async best practices
]
lint.ignore = [
    "E501", # line length handled separately
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ----------------------------------------
# mypy configuration
# ----------------------------------------
[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict = true

# Ignore missing type hints in dependencies, except ours
ignore_missing_imports = true

[tool.setuptools.package-data]
"seymourlib" = ["py.typed"]

# ----------------------------------------
# Pytest configuration
# ----------------------------------------
[tool.pytest.ini_options]
addopts = "-q --timeout=10"
asyncio_mode = "auto"
testpaths = ["tests"]
